feat: change creation structure
